refactor: change svg use
