+++ package.json
@@ -14,6 +14,7 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
+    "react-svg": "^15.1.3",
     "typescript": "^4.7.4",
     "web-vitals": "^2.1.4"
   },
