+++ src/components/user/UserTopBar.tsx
@@ -1,15 +1,10 @@
 import '../../styles/profile.css';
+import { ReactComponent as BackIcon } from '../../assets/back.svg';
 
-// only placeholders for now, gotta change them soon
 export default function UserTopBar() {
   return (
     <div className="user-top-navbar">
-      <svg viewBox="0 0 24 24" aria-hidden="true" className="svg-smaller">
-        <g>
-          <path d="M20 11H7.414l4.293-4.293c.39-.39.39-1.023 0-1.414s-1.023-.39-1.414 0l-6 6c-.39.39-.39 1.023 0 1.414l6 6c.195.195.45.293.707.293s.512-.098.707-.293c.39-.39.39-1.023 0-1.414L7.414 13H20c.553 0 1-.447 1-1s-.447-1-1-1z"></path>
-        </g>
-      </svg>
-
+      <BackIcon className="svg-smaller" />
       <div className="flex flex-col">
         <p className="font-semibold text-white text-2xl user-profile-username">
           user
