+++ src/styles/home.css
@@ -41,6 +41,7 @@
 
 
 .svg-small {
+    cursor: pointer;
     width: 27px;
     filter: invert(100%) sepia(0%) saturate(7500%) hue-rotate(143deg) brightness(104%) contrast(100%);
 }
